requirements.txt
djangorestframework 